requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
openpyxl
google-genai
python-dotenv